Fixed small merge conflict error and re-organized files
